+++ src/features/Update.hpp
@@ -2,4 +2,8 @@
 
 namespace Update {
     void draw(LevelEditorLayer* editor);
+
+    void updateLayerAlpha(GameObject* obj);
+    void updateColorCache();
+    void updateThicknessCache();
 }