+++ src/utils/Enums.hpp
@@ -1,0 +1,8 @@
+#pragma once
+
+enum class MoveIndicatorMode {
+    OnTrigger,
+    Rect,
+    Cluster,
+    Disable
+};