+++ src/utils/Types.hpp
@@ -1,6 +1,6 @@
 #pragma once
 
-// idk this is prolly better than misusing ccrect
+// every so slightly more convenient than ccrect
 struct CornerRect {
     cocos2d::CCPoint p1;
     cocos2d::CCPoint p2;
@@ -9,5 +9,10 @@ struct CornerRect {
     
     CornerRect(cocos2d::CCPoint c1, cocos2d::CCPoint c2) {
         p1 = c1; p2 = c2;
+    }
+
+    CornerRect& operator+=(cocos2d::CCPoint p) {
+        p1 += p; p2 += p;
+        return *this;
     }
 };